Feito o try e catch quando o usuário for adicionar um novo produto e feito algumas validações de formatação

+++ Sistema de Mercado/Produto.cs
@@ -8,7 +8,7 @@ namespace Sistema_de_Mercado
         public int Id { get; set; }
         public string? Nome { get; set; }
         public string? Marca { get; set; }
-        public int CodigoBarras { get; set; }
+        public long CodigoBarras { get; set; }
         public DateTime DataVencimento { get; set; }
         public DateTime DataCadastro { get; set; }
